% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabin.R
\name{FABIN}
\alias{FABIN}
\title{Function to generate factor loadings and standard errors from a vectorized covariance matrix having factor indicators and model-implied instrumental variables}
\usage{
FABIN(tidy_Xs, tidy_Ys, tidy_MIIVs, indicators, my_cov)
}
\arguments{
\item{tidy_Xs}{tidy_Xs, a vectorized covariance matrix of factor indicator variables, generated by \code{\link{tidy_XY}}}

\item{tidy_Ys}{a vectorized covariance matrix of model-implied instrumental variables, generated by \code{\link{tidy_XY}}}

\item{tidy_MIIVs}{character vector naming model-implied instrumental variables, generated by \code{\link{tidy_XY}}}

\item{indicators}{names of variables assumed to have loadings of 1 on the latent factors (character vector)}

\item{my_cov}{covariance matrix to be factor-analyzed, must have row and column names}
}
\value{
results frame, a data frame containing factor loadings and their heteroscedasticity consistent standard errors
}
\description{
Function to generate factor loadings and standard errors from a vectorized covariance matrix having factor indicators and model-implied instrumental variables
}
