% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabin.R
\name{IV_phi}
\alias{IV_phi}
\title{Estimate factor variance, or factor variance-covariance matrix, \eqn{\Phi}}
\usage{
IV_phi(my_cov, my_loadings, my_psi)
}
\arguments{
\item{my_loadings}{named numeric matrix containing factor loadings}

\item{my_psi}{diagonal matrix containing estimated factor unique variances}
}
\value{
phi, factor variance-covariance matrix
}
\description{
Estimate factor variance, or factor variance-covariance matrix, \eqn{\Phi}
}
