% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabin.R
\name{IV_psi}
\alias{IV_psi}
\title{Calculate unique variances given a set of factor loadings and a sample covariance matrix, \eqn{\Psi}}
\usage{
IV_psi(my_cov, my_loadings)
}
\arguments{
\item{my_loadings}{named numeric matrix containing factor loadings}
}
\value{
psi, a named diagonal matrix containing item unique variance estimates
}
\description{
Calculate unique variances given a set of factor loadings and a sample covariance matrix, \eqn{\Psi}
}
