% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabin.R
\name{tidy_XY}
\alias{tidy_XY}
\title{Generates two vectorizations of a covariance matrix, each with row and column names stored explicitly as character vectors.
Vectorizations are to be used as predictor and outcome variables in a regression to estimate factor loadings and standard errors
Can apply weighting for application of FABIN3 or meta-analysis}
\usage{
tidy_XY(my_cov, indicators, weighting = "NEITHER")
}
\arguments{
\item{my_cov}{covariance matrix to be factor-analyzed, must have row and column names}

\item{indicators}{names of variables assumed to have loadings of 1 on the latent factors (character vector)}

\item{weighting}{One of "FABIN3", "meta", or neither: reweight the vectorized covariance matrix for FABIN3 or in a meta-analysis?}
}
\value{
a list containing tidy_Xs, a vectorized covariance matrix of factor indicator variables,  tidy_Ys, a vectorized covariance matrix of model-implied instrumental variables, and tidy_MIIVs, a character vector naming model-implied instrumental variables.
}
\description{
Generates two vectorizations of a covariance matrix, each with row and column names stored explicitly as character vectors.
Vectorizations are to be used as predictor and outcome variables in a regression to estimate factor loadings and standard errors
Can apply weighting for application of FABIN3 or meta-analysis
}
